+++ CHANGELOG.md
@@ -43,6 +43,7 @@ All notable changes to this project will be documented in this file.
 - Do not notify user if the backups are triggered from cronjob.
 - Fixed issue with suspending resources when logged directly as a normal user.
 - Fixed issue with unsuspending a user, PHP-FPM website pool configuration was being deleted.
+- Fixed issue where users created via v-restore-user script were not added to hestia-users group.
 
 ## [1.0.6] - 2019-09-24 - Hotfix
 ### Bugfixes